+++ docx4j-core/src/main/java/org/docx4j/openpackaging/contenttype/ContentTypeManager.java
@@ -449,7 +449,9 @@ public class ContentTypeManager  {
 		} else if (contentType.equals(ContentTypes.IMAGE_WMF)) {
 			return new MetafileWmfPart(new PartName(partName));
 		} else if (contentType.equals(ContentTypes.IMAGE_WEBP)) {
-			return new org.docx4j.openpackaging.parts.WordprocessingML.ImageWebPPart(new PartName(partName));			
+			return new org.docx4j.openpackaging.parts.WordprocessingML.ImageWebPPart(new PartName(partName));
+		} else if (contentType.equals(ContentTypes.IMAGE_JPEG_XR)) {
+			return new org.docx4j.openpackaging.parts.WordprocessingML.ImageJpegXrPart(new PartName(partName));
 		} else if (contentType.equals(ContentTypes.WORDPROCESSINGML_PEOPLE)) {
 			return new PeoplePart(new PartName(partName));
 		} else if (contentType.equals(ContentTypes.VML_DRAWING)) {